Rename syntactic -> syntax
